Cleaned up a bunch of things.

+++ src/dg_variables.cpp
@@ -705,7 +705,7 @@ in the vault (vnum: 453) now and then. you can just use
           }
           if (!strcasecmp(field, "exp")) {
             if (subfield && *subfield) {
-              int64_t addition = MIN(atoi(subfield), 2100000000);
+              int64_t addition = MIN(atoll(subfield), 2100000000);
 
               gain_exp(c, addition);
             }
@@ -848,7 +848,7 @@ in the vault (vnum: 453) now and then. you can just use
         case 'm':
           if (!strcasecmp(field, "maxhitp")) {
             if (subfield && *subfield) {
-              int64_t addition = atol(subfield);
+              int64_t addition = atoll(subfield);
               GET_MAX_HIT(c) = MAX(GET_MAX_HIT(c) + addition, 1);
             }
             snprintf(str, slen, "%" I64T "", GET_MAX_HIT(c));
@@ -856,14 +856,14 @@ in the vault (vnum: 453) now and then. you can just use
 
           else if (!strcasecmp(field, "mana")) {
             if (subfield && *subfield) {
-              int64_t addition = atol(subfield);
+              int64_t addition = atoll(subfield);
               GET_MANA(c) += addition;
             }
             snprintf(str, slen, "%" I64T "", GET_MANA(c));
           }
           else if (!strcasecmp(field, "maxmana")) {
             if (subfield && *subfield) {
-              int64_t addition = atol(subfield);
+              int64_t addition = atoll(subfield);
               GET_MAX_MANA(c) = MAX(GET_MAX_MANA(c) + addition, 1);
             }
             snprintf(str, slen, "%" I64T "", GET_MAX_MANA(c));
@@ -871,7 +871,7 @@ in the vault (vnum: 453) now and then. you can just use
 
           else if (!strcasecmp(field, "move")) {
             if (subfield && *subfield) {
-              int64_t addition = atol(subfield);
+              int64_t addition = atoll(subfield);
               GET_MOVE(c) += addition;
             }
             snprintf(str, slen, "%" I64T "", GET_MOVE(c));
@@ -879,7 +879,7 @@ in the vault (vnum: 453) now and then. you can just use
 
           else if (!strcasecmp(field, "maxmove")) {
             if (subfield && *subfield) {
-              int64_t addition = atol(subfield);
+              int64_t addition = atoll(subfield);
               GET_MAX_MOVE(c) = MAX(GET_MAX_MOVE(c) + addition, 1);
             }
             snprintf(str, slen, "%" I64T "", GET_MAX_MOVE(c));
